style: phpcsfixer

+++ src/Http/Api/Controller/CommentsController.php
@@ -10,6 +10,7 @@ use App\Domain\Comment\Entity\Comment;
 use App\Http\Api\Resource\CommentResource;
 use App\Http\Controller\AbstractController;
 use App\Http\Security\CommentVoter;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Attribute\MapQueryParameter;
@@ -21,27 +22,24 @@ use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 class CommentsController extends AbstractController
 {
-
-
-    #[Route("/comments", name: "comments", methods: ['GET'])]
+    #[Route('/comments', name: 'comments', methods: ['GET'])]
     public function index(
         #[MapQueryParameter(name: 'content')]
         int $contentId,
         CommentRepository $commentRepository,
-    )
-    {
+    ): JsonResponse {
         $comments = $commentRepository->findForApi($contentId);
-        return $this->json(array_map(fn(Comment $comment) => CommentResource::fromComment($comment), $comments), context: ['groups' => ['read']]);
+
+        return $this->json(array_map(fn (Comment $comment) => CommentResource::fromComment($comment), $comments), context: ['groups' => ['read']]);
     }
 
-    #[Route("/comments", methods: ['POST'])]
+    #[Route('/comments', methods: ['POST'])]
     public function create(
         Request $request,
         SerializerInterface $serializer,
         ValidatorInterface $validator,
-        CommentService $service
-    )
-    {
+        CommentService $service,
+    ): JsonResponse {
         $user = $this->getUser();
         $groups = $user === null ? ['anonymous', 'write'] : ['write'];
         $data = $serializer->deserialize(
@@ -51,7 +49,7 @@ class CommentsController extends AbstractController
             ['groups' => $groups]
         );
         if (!($data instanceof CreateCommentDTO)) {
-            throw new \RuntimeException('Expected CreateCommentDTO got ' . $data::class);
+            throw new \RuntimeException('Expected CreateCommentDTO got '.$data::class);
         }
 
         $errors = $validator->validate($data, groups: $groups);
@@ -60,31 +58,33 @@ class CommentsController extends AbstractController
         }
 
         $comment = $service->create($data);
+
         return $this->json(CommentResource::fromComment($comment), context: ['groups' => ['read']]);
     }
 
-    #[Route("/comments/{comment}", name: "comment", methods: ['DELETE'])]
+    #[Route('/comments/{comment}', name: 'comment', methods: ['DELETE'])]
     #[IsGranted(CommentVoter::DELETE, 'comment')]
     public function delete(
         Comment $comment,
-        CommentService $service
-    )
-    {
-        $service->delete($comment->getId());
-        return $this->json(CommentResource::fromComment($comment), context: ['groups' => ['read']]);
+        CommentService $service,
+    ): JsonResponse {
+        if ($comment->getId()) {
+            $service->delete($comment->getId());
+        }
+
+        return new JsonResponse(null, Response::HTTP_NO_CONTENT);
     }
 
-    #[Route("/comments/{comment}", methods: ['PUT'])]
+    #[Route('/comments/{comment}', methods: ['PUT'])]
     #[IsGranted(CommentVoter::UPDATE, 'comment')]
     public function update(
-        #[MapRequestPayload(validationGroups: ['write'], serializationContext: ['write'])]
+        #[MapRequestPayload(serializationContext: ['groups' => ['write']], validationGroups: ['write'])]
         UpdateCommentDTO $data,
         Comment $comment,
         CommentService $service,
-    )
-    {
+    ): JsonResponse {
         $service->update($comment, $data);
+
         return $this->json(CommentResource::fromComment($comment), context: ['groups' => ['read']]);
     }
-
 }
